Tracking v1.1 branch to dev all complete

+++ lib/presentation/blocs/auth/login/login_bloc.dart
@@ -1,5 +1,6 @@
 import 'package:bloc/bloc.dart';
 import 'package:equatable/equatable.dart';
+import 'package:firebase_messaging/firebase_messaging.dart';
 import 'package:tracking_apps/common/shared_preferance_service.dart';
 import 'package:tracking_apps/configs/network/http_response_model.dart';
 import 'package:tracking_apps/configs/network/user_service.dart';
@@ -24,8 +25,17 @@ class LoginBloc extends Bloc<LoginEvent, LoginState> {
 
           if (validateResponse.data != null) {
             await userService.saveAuthTokenToSP(loginResponse.data);
-            await SharedPreferencesService.instance.setData<String>(PreferenceKey.userRole, loginResponse.data);
+            await SharedPreferencesService.instance
+                .setData<String>(PreferenceKey.userRole, loginResponse.data);
             final user = UserModel.fromMap(validateResponse.data);
+
+            String? deviceToken = await FirebaseMessaging.instance.getToken();
+            if (deviceToken != null) {
+              await userService.updateDeviceToken(
+                authToken: loginResponse.data,
+                deviceToken: deviceToken,
+              );
+            }
             emit(LoginSuccess(
                 user: user,
                 message: validateResponse.message,
@@ -84,10 +94,46 @@ class LoginBloc extends Bloc<LoginEvent, LoginState> {
     });
 
     on<LogoutButtonPressed>((event, emit) async {
-      await userService.deleteAuthTokenFromSP();
       emit(const LoginState(isLoading: false));
-    });
+      try {
+        final token = await userService.getAuthTokenFromSP();
+        if (token == null) {
+          emit(const LoginState(
+              message: 'User is not authenticated. Please log in.',
+              isLoading: false,
+              statusCode: 401));
+          print('================== AUTH TOKEN IS NULL! ==================');
+          return;
+        }
 
+        print('================== SENDING LOGOUT REQUEST ==================');
+        print('User Logout with Token : ${event.authToken}');
+
+        HttpResponseModel<dynamic> logoutResponse =
+            await userService.delete(authToken: event.authToken);
+        print('Logout Response : ${logoutResponse.message}');
+
+        if (logoutResponse.statusCode == 200) {
+          await userService.deleteAuthTokenFromSP();
+          emit(LogoutSuccess(
+            message: logoutResponse.message,
+            isLoading: false,
+            statusCode: logoutResponse.statusCode,
+          ));
+          print('================== IN LOGOUT SUCCESS BLOC ==================');
+        } else {
+          emit(LogoutFailed(
+            message: logoutResponse.message,
+            isLoading: false,
+            statusCode: logoutResponse.statusCode,
+          ));
+        }
+      } catch (e) {
+        emit(LogoutFailed(
+            message: e.toString(), isLoading: false, statusCode: 500));
+        print('================== IN LOGOUT FAILED BLOC ==================');
+      }
+    });
 
     on<ClearLoginData>((event, emit) async {
       emit(const LoginState());
